Continue project refactor
