+++ src/components/word/style.css
