+++ src/utils/landing-lyrics.js
